+++ lib/grape/validations/validators/all_or_none.rb
@@ -4,7 +4,7 @@ module Grape
     class AllOrNoneOfValidator < MultipleParamsBase
       def validate!(params)
         super
-        if scope_requires_params && only_subset_present
+        if scope_requires_params && only_subset_present(params)
           raise Grape::Exceptions::Validation, params: all_keys, message: message(:all_or_none)
         end
         params
@@ -12,8 +12,12 @@ module Grape
 
       private
 
-      def only_subset_present
-        scoped_params.any? { |resource_params| !keys_in_common(resource_params).empty? && keys_in_common(resource_params).length < attrs.length }
+      def only_subset_present(params)
+        scoped_params.any? do |resource_params|
+          next unless scope_should_validate?(resource_params, params)
+
+          !keys_in_common(resource_params).empty? && keys_in_common(resource_params).length < attrs.length
+        end
       end
     end
   end