+++ lib/grape/validations/validators/at_least_one_of.rb
@@ -4,7 +4,7 @@ module Grape
     class AtLeastOneOfValidator < MultipleParamsBase
       def validate!(params)
         super
-        if scope_requires_params && no_exclusive_params_are_present
+        if scope_requires_params && no_exclusive_params_are_present(params)
           raise Grape::Exceptions::Validation, params: all_keys, message: message(:at_least_one)
         end
         params
@@ -12,8 +12,12 @@ module Grape
 
       private
 
-      def no_exclusive_params_are_present
-        scoped_params.any? { |resource_params| keys_in_common(resource_params).empty? }
+      def no_exclusive_params_are_present(params)
+        scoped_params.any? do |resource_params|
+          next unless scope_should_validate?(resource_params, params)
+
+          keys_in_common(resource_params).empty?
+        end
       end
     end
   end