+++ lib/grape/validations/validators/default.rb
@@ -20,6 +20,8 @@ module Grape
       def validate!(params)
         attrs = AttributesIterator.new(self, @scope, params)
         attrs.each do |resource_params, attr_name|
+          next unless @scope.meets_dependency?(resource_params, params)
+
           if resource_params.is_a?(Hash) && resource_params[attr_name].nil?
             validate_param!(attr_name, resource_params)
           end