+++ lib/grape/validations/validators/exactly_one_of.rb
@@ -4,7 +4,7 @@ module Grape
     class ExactlyOneOfValidator < MutualExclusionValidator
       def validate!(params)
         super
-        if scope_requires_params && none_of_restricted_params_is_present
+        if scope_requires_params && none_of_restricted_params_is_present(params)
           raise Grape::Exceptions::Validation, params: all_keys, message: message(:exactly_one)
         end
         params
@@ -21,8 +21,12 @@ module Grape
 
       private
 
-      def none_of_restricted_params_is_present
-        scoped_params.any? { |resource_params| keys_in_common(resource_params).empty? }
+      def none_of_restricted_params_is_present(params)
+        scoped_params.any? do |resource_params|
+          next unless scope_should_validate?(resource_params, params)
+
+          keys_in_common(resource_params).empty?
+        end
       end
     end
   end